requirements.txt
pyrogram==2.0.106
tgcrypto
python-dotenv
pillow
apscheduler
